move dict to other file
